Fixed apps integration to npm package system
