+++ apps/cli/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@basilisk/basilisk-cli",
+  "name": "@basilisk/cli",
   "version": "0.1.0",
   "type": "module",
   "dependencies": {
