+++ apps/relay/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@basilisk/basilisk-relay",
+  "name": "@basilisk/relay",
   "version": "0.1.0",
   "type": "module",
   "dependencies": {
