+++ tsconfig.json
@@ -1,7 +1,9 @@
 {
   "files": [],
   "references": [
-    {"path": "packages/core"},
-    {"path": "packages/utils"}
+    { "path": "packages/core" },
+    { "path": "packages/utils" },
+    { "path": "apps/cli" },
+    { "path": "apps/relay" }
   ]
 }
